fix: update docs to follow the newest format
